refactor: RepositoryEntity -> RepositoryResponse 네이밍 수정
