+++ app/src/main/java/nextstep/github/data/model/response/RepositoryResponse.kt
@@ -1,10 +1,10 @@
-package nextstep.github.data.entity
+package nextstep.github.data.model.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RepositoryEntity(
+data class RepositoryResponse(
     @SerialName("full_name") val fullName: String?,
     @SerialName("description") val description: String?,
 )
