+++ app/src/main/java/nextstep/github/data/service/GithubService.kt
@@ -1,6 +1,6 @@
 package nextstep.github.data.service
 
-import nextstep.github.data.entity.RepositoryEntity
+import nextstep.github.data.model.response.RepositoryResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
 
@@ -9,5 +9,5 @@ interface GithubService {
     @GET("orgs/{organization}/repos")
     suspend fun getRepositories(
         @Path("organization") organization: String
-    ): List<RepositoryEntity>
+    ): List<RepositoryResponse>
 }
