+++ app/src/main/java/nextstep/github/ui/model/RepositoryModel.kt
@@ -1,13 +1,13 @@
 package nextstep.github.ui.model
 
-import nextstep.github.data.entity.RepositoryEntity
+import nextstep.github.data.model.response.RepositoryResponse
 
 data class RepositoryModel(
     val fullName: String,
     val description: String
 )
 
-fun RepositoryEntity.toModel(): RepositoryModel = RepositoryModel(
+fun RepositoryResponse.toModel(): RepositoryModel = RepositoryModel(
     fullName = fullName ?: "",
     description = description ?: ""
 )
